src/requirements.txt
auto_mix_prep==0.2.0
colorlog==6.9.0
fastapi==0.116.1
playwright==1.53.0
psutil==7.0.0
pydantic==2.11.7
python-dotenv==1.1.1
python-telegram-bot==21.3
Requests==2.32.4
